[Fix] 새로운 기능 추가함에 따라 seed파일 수정
